tools/requirements.txt
pylink_square==1.1.0
rich==13.7.1
